+++ src/effects/spells/spiritShroud.js
@@ -31,6 +31,7 @@ export async function spiritShroudEffect(document) {
     onlyOnce: false,
     displayTemp: true,
   };
+  // setProperty(effect, "flags.dae.selfTarget", true);
 
   const itemMacroText = await loadMacroFile("spell", "spiritShroud.js");
   setProperty(document, "flags.itemacro", generateItemMacroFlag(document, itemMacroText));