+++ src/effects/spells/spiritualWeapon.js
@@ -6,11 +6,13 @@ export async function spiritualWeaponEffect(document) {
   const itemMacroText = await loadMacroFile("spell", "spiritualWeapon.js");
   document.flags["itemacro"] = generateItemMacroFlag(document, itemMacroText);
   effect.changes.push(generateMacroChange("@item.level"));
+  setProperty(effect, "flags.dae.selfTarget", true);
   document.effects.push(effect);
 
   document.data.damage = { parts: [], versatile: "", value: "" };
   document.data['target']['type'] = "self";
-  document.data.range = { value: null, units: "self", long: null };
+  // document.data.range = { value: null, units: "self", long: null };
+
   document.data.actionType = "other";
 
   return document;